requirements.txt
Pillow          # For image processing
websockets      # For WebSocket communication
mcschematic     # For creating Minecraft schematic files